# doc/api.yaml
openapi: 3.0.3
info:
  version: 'latest'
  title: FoodGen - API
  description: |
    A simple API that creates random food
security:
  - BearerAuth: []
paths:
...
  '/users':
    put:
      tags:
        - User
      summary: Create or update users
      operationId: crupdateUsers
      requestBody:
        description: List of users to be crupdated
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/CreateUser'
      responses:
        '200':
          description: return a list crupdated users.
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  '/users/whoami':
    get:
      tags:
        - User
      summary: Tell you who you are
      operationId: whoami
      responses:
        '200':
          description: return the specific user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  '/users/login':
    post:
      tags:
        - User
      summary: Let user log in the application
      operationId: login
      requestBody:
        description: send user credentials to the server
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Auth'
      responses:
        '200':
          description: Logged in
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
  '/users/signup':
    post:
      tags:
        - User
      summary: Let user sign up the application
      operationId: signup
      requestBody:
        description: send new user credentials to the server
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '200':
          description: Logged in
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
...
  '/users/{username}':
    get:
      tags:
        - User
      summary: Get user by username
      operationId: getUserByUsername
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: return the specific user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
...
components:
  responses:
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
    CreateUser:
      allOf:
        - $ref: '#/components/schemas/Auth'
      type: object
      properties:
        firstname:
          type: string
          example: "Rio"
        lastname:
          type: string
          example: "Camero"
        username:
          type: string
        password:
          type: string
        allergies:
          type: array
          items:
            type: string
        preferences:
          type: array
          items:
            type: string
    User:
      allOf:
        - $ref: '#/components/schemas/CreateUser'
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - USER
            - ADMIN
    Auth:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
...
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error